+++ entity/views.py
@@ -60,6 +60,7 @@ def index(request):
     return render(request, 'list_entities.html', context)
 
 
+@airone_profile
 @http_get
 def create(request):
     user = User.objects.get(id=request.user.id)
@@ -171,6 +172,7 @@ def do_edit(request, entity_id, recv_data):
     })
 
 
+@airone_profile
 @http_post([
     {'name': 'name', 'type': str, 'checker': lambda x: (
         x['name'] and not Entity.objects.filter(name=x['name']).exists()
@@ -230,6 +232,7 @@ def do_create(request, recv_data):
     })
 
 
+@airone_profile
 @http_get
 def export(request):
     user = User.objects.get(id=request.user.id)
@@ -267,6 +270,7 @@ def export(request):
     return get_download_response(output, 'entity.yaml')
 
 
+@airone_profile
 @http_post([])
 @check_permission(Entity, ACLType.Full)
 def do_delete(request, entity_id, recv_data):
@@ -299,6 +303,7 @@ def do_delete(request, entity_id, recv_data):
     return JsonResponse(ret)
 
 
+@airone_profile
 @http_get
 def history(request, entity_id):
     if not Entity.objects.filter(id=entity_id).exists():
@@ -315,6 +320,7 @@ def history(request, entity_id):
     return render(request, 'history_entity.html', context)
 
 
+@airone_profile
 @http_get
 def dashboard(request, entity_id):
     if not Entity.objects.filter(id=entity_id).exists():
@@ -384,6 +390,7 @@ def dashboard(request, entity_id):
     return render(request, 'dashboard_entity.html', context)
 
 
+@airone_profile
 @http_get
 def conf_dashboard(request, entity_id):
     if not Entity.objects.filter(id=entity_id).exists():
@@ -401,6 +408,7 @@ def conf_dashboard(request, entity_id):
     return render(request, 'conf_dashboard_entity.html', context)
 
 
+@airone_profile
 @http_post([
     {'name': 'attrs', 'type': list,
      'checker': lambda x: all([EntityAttr.objects.filter(id=v).exists() for v in x['attrs']])}