+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Fixed
 * Fixed not being redirected to the original URL after login
+* Fixed some request logs not output
 * Fixed the log message was not output to django.log
 
 ## v2.8.0